udpate register

+++ src/main/java/com/iotstar/onlinetest/repositories/AccountDAO.java
@@ -5,7 +5,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface AccountDAO extends JpaRepository<Account, Integer> {
+public interface AccountDAO extends JpaRepository<Account, Long> {
     public Account getByUsername(String username);
