+++ src/main/java/com/iotstar/onlinetest/repositories/UserDAO.java
@@ -5,7 +5,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface UserDAO extends JpaRepository<User, Integer> {
-    public User getUserByUserId(int userId);
+public interface UserDAO extends JpaRepository<User, Long> {
+    public User getUserByUserId(Long userId);
 
 }
